Add infinityScroll & Refactor Component Modularization
