+++ frontend/src/components/Bartering/barteringList.scss
@@ -54,3 +54,7 @@ $gray: #8b8b8b;
     }
   }
 }
+
+.no_item {
+  width: 100%;
+}
