+++ frontend/src/components/Bartering/category.scss
@@ -1,0 +1,5 @@
+.categorySelect {
+  width: 100%;
+  font-size: 1.2rem;
+  padding: 1rem;
+}
