+++ frontend/src/components/Chat/chat.scss
@@ -14,6 +14,14 @@
   background-color: white;
   box-shadow: 0 1px 1px 0 rgba(0.5, 0.5, 0.5, 0.5);
 }
+.chat-header-close {
+  position: fixed;
+  top: 0;
+  right: 0;
+  padding: 1rem 1.5rem;
+  z-index: 1001;
+  font-size: 2rem;
+}
 .chat-section {
   font-family: "Helvetica Neue";
   font-size: 1.2rem;
