+++ frontend/src/components/Detail/body.scss
