+++ frontend/src/components/Detail/footer.scss
