+++ frontend/src/components/Write&Update/body.scss
