+++ frontend/src/components/보일러.tsx
@@ -1,0 +1,9 @@
+import React, { useEffect, useState, useRef } from "react";
+
+export const BarteringWrapper = (props: any) => {
+  return (
+    <>
+      <div>dd</div>
+    </>
+  );
+};
