+++ frontend/src/router/index.ts
@@ -1,11 +1,10 @@
-export { default as Bartering } from "../views/Bartering";
-export { default as BarteringWrite } from "../views/BarteringWrite";
-export { default as BarteringDetail } from "../views/BarteringDetail";
-export { default as Signin } from "../views/Signin";
-export { default as Signup } from "../views/Signup";
-export { default as Test } from "../views/Test";
-export { default as Info } from "../views/Info";
-export { default as ChatRoom } from "../views/ChatRoom";
-export { default as Chat } from "../views/Chat";
-export { default as Loading } from "../views/Loading";
-export { default as BarteringUpdate } from "../views/BarteringUpdate";
+export { default as Bartering } from "views/Bartering";
+export { default as BarteringWrite } from "views/BarteringWrite";
+export { default as BarteringDetail } from "views/BarteringDetail";
+export { default as Signin } from "views/Signin";
+export { default as Signup } from "views/Signup";
+export { default as Info } from "views/Info";
+export { default as ChatRoom } from "views/ChatRoom";
+export { default as Chat } from "views/Chat";
+export { default as Loading } from "views/Loading";
+export { default as BarteringUpdate } from "views/BarteringUpdate";
