--- frontend/src/style/test.scss
@@ -1,44 +1,0 @@
-.test-content {
-  width: 100%;
-  div {
-    padding: 20px 50px;
-    &:nth-child(1) {
-      padding-top: 50px;
-    }
-    .test-title {
-      font-family: "S-CoreDream-9Black";
-      display: inline;
-      border-bottom: 5px solid black;
-      &::after {
-        content: "";
-        display: block;
-        width: 100%;
-        height: 1px;
-        margin-bottom: 10px;
-      }
-    }
-    .test-description {
-      font-family: "BMJUA";
-      width: 100%;
-      display: flex;
-      justify-content: center;
-      text-align: center;
-      padding: 25px;
-      font-size: large;
-      ul {
-        list-style: none;
-        margin: 0;
-        padding: 0;
-      }
-      li {
-        font-size: large;
-        display: inline-block;
-        padding: 15px;
-        margin: 5px;
-      }
-    }
-    .color {
-      border: solid 1px black;
-    }
-  }
-}
