+++ frontend/src/views/BarteringWrite.tsx
@@ -1,53 +1,28 @@
-import React, { useState, ChangeEvent, useEffect } from "react";
-import Nav from "../components/Nav";
-import http from "../api/http-common";
-import { useHistory } from "react-router-dom";
+import React, { useState, useEffect } from "react";
+import http from "api/http-common";
 import jwt_decode from "jwt-decode";
 
-import "../style/barteringwrite.scss";
+import { Header } from "components/Write&Update/Header";
+import { useHistory } from "react-router-dom";
+import { PhotoSelecter } from "components/Write&Update/PhotoSelecter";
+import { ImagePreview } from "components/Write&Update/IamgePreview";
+import { UpdateContent } from "components/Write&Update/Content";
+import { Nav } from "components/Nav";
+
+import "components/Write&Update/body.scss";
 
 const BarteringWrite = () => {
+  let history = useHistory();
   const [imagePreviewList, setImagePreviewList] = useState<Array<string>>([]);
   const [title, setTitle] = useState("");
   const [category, setCategory] = useState("");
-  const [categoryList, setCategoryList] = useState<Array<string>>([]);
   const [contents, setContents] = useState("");
   const [need, setNeed] = useState("");
   const [uploadFileList, setUploadFileList] = useState<Array<File>>([]);
-  let history = useHistory();
 
-  const loadFile = (e: any) => {
-    const files = e.target.files;
-    let image: Array<string> = [...imagePreviewList];
-    let file: Array<File> = [...uploadFileList];
-
-    if (image.length !== 5) {
-      Array.from(files).forEach((v: any) => {
-        if (image.length < 5) {
-          image.push(URL.createObjectURL(v));
-          file.push(v);
-        }
-      });
-    }
-    setImagePreviewList(image);
-    setUploadFileList(file);
-  };
-
-  const onChangeHandle = (
-    e: ChangeEvent<HTMLInputElement | HTMLSelectElement | HTMLTextAreaElement>
-  ) => {
-    if (e.target.className === "title") {
-      setTitle(e.target.value);
-    } else if (e.target.className === "categorySelect") {
-      setCategory(e.target.value);
-    } else if (e.target.className === "contents") {
-      setContents(e.target.value);
-    } else if (e.target.className === "need") {
-      setNeed(e.target.value);
-    } else {
-      return;
-    }
-  };
+  useEffect(() => {
+    localStorage.setItem("nowPath", "/BarteringWrite");
+  }, []);
 
   const onSubmitHandle = (e: any) => {
     e.preventDefault();
@@ -59,7 +34,6 @@ const BarteringWrite = () => {
     }
 
     let formData = new FormData();
-
     formData.append("userId", decodedToken.id);
     formData.append("title", title);
     formData.append("category", category);
@@ -67,28 +41,14 @@ const BarteringWrite = () => {
     formData.append("address", decodedToken.address);
     formData.append("need", need);
 
-    // console.log(uploadFileList);
-
     uploadFileList.forEach(file => {
       formData.append("images", file);
     });
 
-    console.log(
-      "userID : ",
-      String(decodedToken.id),
-      "\ntitle : ",
-      title + "\ncatetory : ",
-      category,
-      "\naddress : ",
-      decodedToken.address,
-      "\nneed : ",
-      need,
-      "\ncontents : ",
-      contents,
-      "\nfile : ",
-      uploadFileList
-    );
+    getWriteItemAPI(formData);
+  };
 
+  const getWriteItemAPI = (formData: FormData) => {
     http
       .post("/item", formData, {
         headers: {
@@ -96,133 +56,57 @@ const BarteringWrite = () => {
           Accept: "application/json"
         }
       })
-      .then(({ data }) => {
+      .then(() => {
         alert("등록완료!");
-        console.log(data);
+        history.push("/Home");
       })
-      .catch(error => {
-        console.log(error);
+      .catch(e => {
+        console.log(e);
       });
   };
 
-  useEffect(() => {
-    localStorage.setItem("nowPath", "/BarteringWrite");
-
-    http
-      .get("/category")
-      .then(({ data }) => {
-        setCategoryList(data);
-      })
-      .catch(error => {
-        console.log(error);
-      });
-  }, []);
-
-  const close = () => {
-    history.push("/Home");
+  const onPhotoSelecter = (photoSelecter: any) => {
+    setImagePreviewList(photoSelecter);
   };
+
+  const onFileSelecter = (fileSelecter: any) => {
+    setUploadFileList(fileSelecter);
+  };
+
+  const onTitle = (title: string) => {
+    setTitle(title);
+  };
+
+  const onCategory = (category: string) => {
+    setCategory(category);
+  };
+
+  const onContents = (contents: string) => {
+    setContents(contents);
+  };
+
+  const onNeed = (need: string) => {
+    setNeed(need);
+  };
+
+  const state = { onTitle, onCategory, onContents, onNeed };
+  const item = { title, category, contents, need };
 
   return (
     <>
       <section className="router-container-fixed router-top router-footer">
         <form onSubmit={onSubmitHandle}>
-          <div className="write-header-wrap">
-            <span className="close" onClick={close}>
-              X
-            </span>
-            <span>상 품 등 록</span>
-            <input type="submit" value="완료" />
-          </div>
+          <Header title={"상 품 등 록"} />
           <article className="article-area">
             <div className="write-body-wrap">
-              <label>
-                <figure>
-                  <img
-                    src={require("../assets/icon/photo.png")}
-                    className="photoImage"
-                    alt="photoImage"
-                  />
-                  <figcaption>{imagePreviewList.length}/5</figcaption>
-                </figure>
-                <input
-                  type="file"
-                  onChange={loadFile}
-                  accept=".gif, .jpg, .png, .jpeg"
-                  multiple
-                />
-              </label>
-              <div className="bartering-image-content">
-                {imagePreviewList &&
-                  imagePreviewList.map((v, i) => {
-                    return (
-                      <>
-                        <img
-                          key={v}
-                          className="PhotoResult"
-                          src={v}
-                          alt="img"
-                          // ref={el => (el === null ? el : (imgRef.current[i] = el))}
-                        />
-                      </>
-                    );
-                  })}
-              </div>
+              <PhotoSelecter
+                onPhotoSelecter={onPhotoSelecter}
+                onFileSelecter={onFileSelecter}
+                item={imagePreviewList}
+              />
+              <ImagePreview previewList={imagePreviewList} />
             </div>
-            <div className="write-body-wrap-column">
-              <div>
-                <label className="title">상품 이름</label>
-                <input
-                  className="title"
-                  type="text"
-                  placeholder="상품 이름"
-                  value={title}
-                  onChange={onChangeHandle}
-                  required
-                />
-              </div>
-
-              <div>
-                <label className="category">카테고리</label>
-                <select
-                  className="categorySelect"
-                  value={category}
-                  onChange={onChangeHandle}
-                >
-                  <option value="">카테고리 선택</option>
-
-                  {categoryList.map(v => {
-                    return (
-                      <option key={v} value={v}>
-                        {v}
-                      </option>
-                    );
-                  })}
-                </select>
-              </div>
-
-              <div>
-                <label className="need">교환 받을 물건 / 필요한 물건</label>
-                <input
-                  className="need"
-                  type="text"
-                  placeholder="교환 받을 물건 / 필요한 물건"
-                  value={need}
-                  required
-                  onChange={onChangeHandle}
-                />
-              </div>
-
-              <div>
-                <label>상품 설명</label>
-                <textarea
-                  placeholder="상품에 대한 설명을 작성해주세요."
-                  className="contents"
-                  value={contents}
-                  required
-                  onChange={onChangeHandle}
-                ></textarea>
-              </div>
-            </div>
+            <UpdateContent state={state} item={item} />
           </article>
         </form>
       </section>
