+++ frontend/src/views/Signin.tsx
@@ -1,12 +1,13 @@
-import React, { useState, ChangeEvent, useEffect, useRef } from "react";
+import React, { useState, useEffect } from "react";
+import http from "api/http-common";
+
 import { Link, useHistory } from "react-router-dom";
-import { setToken, setAuth, setPath } from "../redux/action";
-import { AuthState, PathState } from "../redux/reducer";
+import { setToken, setAuth } from "redux/action";
+import { AuthState } from "redux/reducer";
 import { useSelector, useDispatch } from "react-redux";
+import { SigninContent } from "components/User/SigninContent";
 
-import http from "../api/http-common";
-
-import "../style/user.scss";
+import "components/User/body.scss";
 
 const Signin = () => {
   const [email, setEmail] = useState("");
@@ -15,20 +16,12 @@ const Signin = () => {
   let history = useHistory();
   const dispatch = useDispatch();
   const auth = useSelector<AuthState, AuthState["auth"]>(state => state.auth);
-  const path = useSelector<PathState, PathState["path"]>(state => state.path);
-
-  const stayRef = useRef<HTMLInputElement>(null);
-  const storageRef = useRef<HTMLInputElement>(null);
 
   useEffect(() => {
     if (auth && localStorage.getItem("stayLogin")) {
-      history.push(path);
+      history.push(String(localStorage.getItem("nowPath")));
     }
-    if (localStorage.getItem("email") && storageRef.current) {
-      setEmail(String(localStorage.getItem("email")));
-      storageRef.current.checked = true;
-    }
-  }, [auth, history, path]);
+  }, [auth]);
 
   const handleSubmit = (e: any): void => {
     e.preventDefault();
@@ -39,45 +32,36 @@ const Signin = () => {
       })
       .then(({ data }) => {
         dispatch(setToken(data));
-        dispatch(setPath("/Home"));
         dispatch(setAuth(true));
         localStorage.setItem("token", data);
 
         history.push("/Loading");
       })
       .catch(e => {
-        alert(e);
+        console.log(e);
         dispatch(setToken(""));
         dispatch(setAuth(false));
-        console.log(e.response.data);
+        if (e.request.status === 401) {
+          alert("존재하지 않은 아이디 혹은 비밀번호가 틀립니다.");
+        }
       });
   };
 
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    if (e.target.type === "email") {
-      if (storageRef.current && storageRef.current.checked) {
-        localStorage.setItem("email", e.target.value);
-      }
-      setEmail(e.target.value);
-    } else {
-      setPassword(e.target.value);
+  const setState = (data: any, category: string) => {
+    switch (category) {
+      case "email":
+        setEmail(data);
+        break;
+      case "password":
+        setPassword(data);
+        break;
+
+      default:
+        break;
     }
   };
 
-  const onStayLoginHandle = () => {
-    if (stayRef.current && stayRef.current.checked) {
-      localStorage.setItem("stayLogin", "true");
-    } else {
-      localStorage.removeItem("stayLogin");
-    }
-  };
-
-  const onStorageIDHandle = () => {
-    if (storageRef.current && !storageRef.current.checked) {
-      localStorage.removeItem("email");
-    }
-  };
-
+  const item = { email, password };
   return (
     <>
       <section className="user-container">
@@ -90,46 +74,8 @@ const Signin = () => {
               <h2 className="user-title">Unithis</h2>
             </div>
             <form onSubmit={handleSubmit}>
-              <div className="user-input-area">
-                <label>이메일</label>
-                <input
-                  className="email"
-                  type="email"
-                  required
-                  value={email}
-                  onChange={handleChange}
-                />
-              </div>
-              <div className="user-input-area">
-                <label>비밀번호</label>
-                <input
-                  className="password"
-                  type="password"
-                  // required
-                  value={password}
-                  onChange={handleChange}
-                />
-              </div>
-              <div className="checkbox-area">
-                <label>
-                  <input
-                    type="checkbox"
-                    value="아이디저장"
-                    ref={storageRef}
-                    onClick={onStorageIDHandle}
-                  />
-                  아이디저장
-                </label>
-                <label>
-                  <input
-                    type="checkbox"
-                    value="로그인상태유지"
-                    ref={stayRef}
-                    onClick={onStayLoginHandle}
-                  />
-                  로그인상태유지
-                </label>
-              </div>
+              <SigninContent setState={setState} item={item} />
+
               <div className="button-area">
                 <input type="submit" className="btn blue" value="로그인" />
               </div>
